+++ package.json
@@ -13,16 +13,17 @@
     "test": "mocha"
   },
   "dependencies": {
-    "chalk": "^4.1.1",
+    "chalk": "^5.3.0",
     "clear": "^0.1.0",
     "cryptify": "^4.1.2",
-    "figlet": "^1.5.0",
-    "inquirer": "^8.0.0",
-    "minimist": "^1.2.5",
+    "figlet": "^1.7.0",
+    "inquirer": "^9.2.14",
+    "minimist": "^1.2.8",
     "typescript": "^5.3.3"
   },
   "bin": "./index.js",
   "devDependencies": {
+    "@types/chalk": "^2.2.0",
     "@types/node": "^20.11.16",
     "chai": "^4.3.4",
     "eslint": "^7.26.0",
