+++ src/commands/edit.ts
@@ -6,7 +6,7 @@ const clear = require("clear");
 const Cryptify = require("cryptify");
 const inquirer = require("inquirer");
 
-const deleteCmd = (argv) => {
+const edit = (argv) => {
   const CONFIG_DIR_PATH = path.join(process.cwd(), argv.p ? argv.p : "config");
   const ENCRYPTION_KEY_PATH = path.join(CONFIG_DIR_PATH, `${argv.e}.key`);
   const ENCRYPTED_FILE_PATH = path.join(CONFIG_DIR_PATH, `${argv.e}.yml.enc`);
@@ -59,51 +59,57 @@ const deleteCmd = (argv) => {
 
     if (Object.keys(parsedObj).length === 0) {
       return console.log(
-        chalk.red("Nothing to delete. Please add some keys first.")
+        chalk.red("Nothing to edit. Please add some keys first.")
       );
     }
 
     inquirer
       .prompt([
         {
-          name: "keysToDelete",
-          type: "checkbox",
-          message: "Which key(s) would you like to delete?",
+          name: "keyToEdit",
+          type: "list",
+          message: "Which key would you like to edit?",
           choices: Object.keys(parsedObj),
         },
       ])
-      .then((answers) => {
-        const filteredKeys = Object.keys(parsedObj).filter((key) => {
-          if (answers.keysToDelete.includes(key)) {
-            return false;
-          }
+      .then((listAnswers) => {
+        inquirer
+          .prompt([
+            {
+              name: "keyEditedValue",
+              type: "input",
+              message: "What is the new value of this key?",
+              default: parsedObj[listAnswers.keyToEdit],
+              validate: (value) => {
+                if (value.length) {
+                  return true;
+                }
 
-          return true;
-        });
+                return "Please enter the new value of the key you would like to edit.";
+              },
+            },
+          ])
+          .then((inputAnswers) => {
+            parsedObj[listAnswers.keyToEdit] = inputAnswers.keyEditedValue;
 
-        let newObj = {};
+            const encryptedFileInstance = new Cryptify(
+              ENCRYPTED_FILE_PATH,
+              secretKeyData,
+              null,
+              null,
+              true,
+              true
+            );
 
-        filteredKeys.forEach((key) => {
-          newObj[key] = parsedObj[key];
-        });
+            fs.writeFileSync(ENCRYPTED_FILE_PATH, JSON.stringify(parsedObj));
 
-        const encryptedFileInstance = new Cryptify(
-          ENCRYPTED_FILE_PATH,
-          secretKeyData,
-          null,
-          null,
-          true,
-          true
-        );
-
-        fs.writeFileSync(ENCRYPTED_FILE_PATH, JSON.stringify(newObj));
-
-        encryptedFileInstance.encrypt();
-      })
-      .then(() => {
-        console.log("Done! 🌟");
+            encryptedFileInstance.encrypt();
+          })
+          .then(() => {
+            console.log("Done! 🌟");
+          });
       });
   });
 };
 
-module.exports = deleteCmd;
+export default edit;
