+++ src/index.ts
@@ -6,11 +6,9 @@ import clear from "clear";
 import figlet from "figlet";
 
 import init from "./commands/init";
-
-const init = require("./commands/init");
-const edit = require("./commands/edit");
-const add = require("./commands/add");
-const deleteCmd = require("./commands/delete");
+import edit from "./commands/edit";
+import add from "./commands/add";
+import deleteCmd from "./commands/delete";
 
 clear();
 
@@ -21,7 +19,7 @@ console.log(
 const argv = minimist(process.argv.slice(2));
 
 if (argv._.length === 0) {
-  return console.log(chalk.red("Please enter a valid command"));
+  console.log(chalk.red("Please enter a valid command"));
 }
 
 const availableCommands = {
